Updated module doc strings so they are only a single sentence and relocated `print_time()` function to utils.
